+++ changelog.md
@@ -7,3 +7,12 @@
 * Update product component with mock data.
 * Add product component.
 * Initial commit for shop project.
+
+## Task 2
+* Add highlight directive and apply it to the products component (@HostBinding, @HostListener).
+* Create modules for: Products, Cart, Orders, Core, Shared.
+* Divided cart component on cart-list & cart-item.
+* Add app title to the AppComponent (@ViewChild).
+* Add OnPush change strategy to the ProductComponent.
+* Initial commit for shop project.
+* Use ngClass for add Class to element.